requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
numpy==1.26.4
pillow==10.3.0
matplotlib==3.8.4
torch==2.2.2
torchvision==0.17.2
torchlm==0.1.6.10
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
